Python/requirements.txt
numpy
scipy
Pillow
numba
scikit-image